requirements.txt
streamlit
Pillow
requests
opencv-python
python-dotenv
openai
ultralytics